refactor: getPhoneNumberErrorMessage 유틸함수, useFormattedPhoneNumber 훅 common으로 이동
